road_situation/requirements.txt
django
djangorestframework
ml_collections
torch
totchvision
numpy
scipy